+++ include/pfnet/branch.h
@@ -174,6 +174,7 @@ BOOL BRANCH_is_phase_shifter(Branch* br);
 BOOL BRANCH_is_tap_changer(Branch* br);
 BOOL BRANCH_is_tap_changer_v(Branch* br);
 BOOL BRANCH_is_tap_changer_Q(Branch* br);
+BOOL BRANCH_is_part_of_3_winding_transformer(Branch* br);
 Branch* BRANCH_list_reg_add(Branch* reg_br_list, Branch* br);
 Branch* BRANCH_list_reg_del(Branch* reg_br_list, Branch* br);
 int BRANCH_list_reg_len(Branch* reg_br_list);