+++ src/net/net.c
@@ -2777,6 +2777,18 @@ int NET_get_num_slack_buses(Net* net) {
   return n;
 }
 
+int NET_get_num_star_buses(Net* net) {
+  int i;
+  int n = 0;
+  if (!net)
+    return 0;
+  for (i = 0; i < net->num_buses; i++) {
+    if (BUS_is_star(BUS_array_get(net->bus,i)))
+      n++;
+  }
+  return n;
+}
+
 int NET_get_num_buses_reg_by_gen(Net* net) {
   int i;
   int n = 0;